Add debian platform support to nrpe_basic::service
This change follows the same pattern as the change to model
packages on a per-platform basis.

The service resouce name is held constant and the name
parameter is used on a per platform basis.

In addition, the init script on ubuntu does not implement
a status check so one is explicitly provided to the resource.