Add check_mcollective nrpe command
This does a default check to make sure systems have
registered themselves in the past 300 seconds.  As
part of this change set the check_mcollective.rb script
needs to be deployed.  This requires the plugin
directory to be known, so the logic for this data
has been copied from the nrpe_basic::command
defined resource.

This plugdir setting is a good candidate to improve
with hiera.